Set user.dir when launching Jace.
Jace writes a .java.conf file when the user saves Jace settings
by pressing the Save button in the config panel. By defult,
such file is written in the directory from which Jace was launched.
Add a Java option to change the directory to the user's home directory,
so that the file is written to $HOME/.java.conf.