Use stable version of epple2.
We now use the stable version of Epple2, and leave
the Github source as an option.

We have added Epple-II-Rom (epple2sys) as a resource, so that
epple2 won't warn about missing files on launch.

Note that we have been able to build and run the HEAD. The stable
version does not build on OS X Lion (we haven't tried other platforms
yet).