epple2: add caveat on how to quit from full screen.
I haven’t found any way to exit full-screen mode
(pressing F3 goes full screen, but pressing it
again does nothing). When in full-screen mode, it
is not possible to switch application: basically,
you are locked into Epple ][. The only
way to re-gain control of your computer seems to
be cmd-alt-esc, which terminates the program. Since
this is not obvious, a caveat has been added.