Update PLASMA to commit 7f9975c4.
This is v1.0, but it has not been tagged by the author.

For some reason, Homebrew chokes on make with this release. The
workaround I have found consists in runnning make twice, ignoring the
error of the first run. I have no problems building PLASMA when I run
make manually.