Update formula: Jace v3.0
Also remove ZuluFX cask, because that was needed only to build the
previous version of Jace.

Note that the formula gets the released binary and not the source,
because building from source requires Maven 3.8.8, but Homebrew has
a newer version. A fork of GraalVM is also required, which should be
installed as a resource. I think this is too much trouble.